Remove self-enclosing element from Javadoc comments
This commit removes self-enclosing element from Javadoc comments in the
entire source and replace them with their generic counterpart.

As it turned out, self-enclosing elements (such as '<p/>' and the
likes) are troublesome for automated Javadoc JAR creation (which is
required for JCenter publication). So we have to get rid of them.

Resolves: #1