Remove Retrolambda and its usages within the source
This commit removes Retrolambda alongside with its usages within the
source.

The usage of said dependency has proven to cause problems when combined
with Android Gradle Plugin 1.3.0 (especially when trying to generate
Javadoc JAR). Therefore, for the sake of publishing into JCenter
repository, we would have to get rid of it at least until the issue has
been resolved.

Resolves: #2