Remove illegal characters from Javadoc comments in Permissions
This commit remove the usage of illegal characters (that were mistaken
as HTML syntax) from Javadoc comments found in the Permissions class.

As it turned out, said illegal characters (such as '<', '>', and the
likes) are troublesome for automated Javadoc JAR creation (which is
required for JCenter publication). So we have to get rid of them.