Add convenient method to obtain LayoutInflater in adapters
LayoutInflater is one of the most important thing to have in an
adapter. So why not ease the process and make a getter for it? They
already has a reference to a Context (which is all we need to obtain a
LayoutInflater) and almost any subclass would need one anyway.

Oh and beside the above, this commit also carried out some code
rearrangement in the adapters' code.